Day two solution and refactor

+++ src/main/java/advent/Day6/Day6.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public class Day6 implements Solver {
     @Override
-    public int solve(List<String> input) {
+    public long solve(List<String> input) {
         return 0;
     }
 }
