+++ src/main/java/advent/Solver.java
@@ -3,5 +3,5 @@ package advent;
 import java.util.List;
 
 public interface Solver {
-    int solve(List<String> input);
+    long solve(List<String> input);
 }
